+++ tests/Commands/Events/AbstractEventTestCase.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace AvtoDev\AmqpRabbitManager\Tests\Commands\Events;
+
+use AvtoDev\AmqpRabbitManager\Tests\AbstractTestCase;
+
+abstract class AbstractEventTestCase extends AbstractTestCase
+{
+    //
+}