+++ tests/Commands/Events/QueueCreatingTest.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace AvtoDev\AmqpRabbitManager\Tests\Commands\Events;
+
+/**
+ * @covers \AvtoDev\AmqpRabbitManager\Commands\Events\QueueCreating<extended>
+ */
+class QueueCreatingTest extends AbstractEventTestCase
+{
+    /**
+     * @small
+     *
+     * @return void
+     */
+    public function testWip(): void
+    {
+        self::markTestIncomplete();
+    }
+}