+++ tests/ExchangesFactoryTest.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace AvtoDev\AmqpRabbitManager\Tests;
+
+/**
+ * @covers \AvtoDev\AmqpRabbitManager\ExchangesFactory<extended>
+ */
+class ExchangesFactoryTest extends AbstractTestCase
+{
+    /**
+     * @small
+     *
+     * @return void
+     */
+    public function testWip(): void
+    {
+        self::markTestIncomplete();
+    }
+}