+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## v2.0.0
+
+### Added
+
+- Exchanges factory
+- Events: ...
+
+### Changed
+
+- `setup` section in configuration file and "map" in `rabbit:setup` command
+- `rabbit:setup` command now creates exchanges
+
+### Fixed
+
+- Exception messages
+
 ## v1.1.0
 
 ### Added